requirements.txt
Flask
flask-cors
paho-mqtt
psycopg2
uWSGI
